feat: customer phone-lookup, schema consolidation, dead-code purge
- feat: dedicated GET /admin/customers/lookup?phone= returns a single
  nullable Customer (phone is identity, 0-or-1). POS name-prefill uses it
  instead of misusing the paginated browse list. UX-only — checkout still
  find-or-creates by phone server-side (ADR-0011).
- refactor: single-source schemas. Module *.schema.ts own the POST schemas;
  src/schema/index.ts re-exports them for @fresclean/api/schema. Removes the
  server-vs-web divergence (server validated via createInsertSchema while web
  used hand-rolled barrel copies) for customer/store/category/service/
  payment/product. Server now validates creates with the richer schemas the
  web already enforced.
- refactor: shared isUniqueViolation (Postgres 23505) in utils/errors.ts;
  customers + shifts services share it instead of hand-rolling the check.
- chore: purge dead code — orphaned files (form-errors, use-is-desktop), the
  daily-report and my-order-services client chains, pickup-state status
  helpers, buildCategoryTabs, updateOrderTotal, unused inferred types, plus
  cascade orphans (eq import, daysBetween, CategoryTab, OrderFulfillmentLike).
- chore: drop dead buildStoreWhereClause from utils/authorization.ts.
