+++ packages/server/src/modules/categories/category.schema.ts
@@ -1,8 +1,13 @@
-import { createInsertSchema, createUpdateSchema } from "drizzle-orm/zod";
+import { createUpdateSchema } from "drizzle-orm/zod";
 import { z } from "zod";
 import { categoriesTable } from "@/db/schema";
+import { isActiveSchema, textSchema, varcharSchema } from "@/schema/common";
 
-export const POSTCategorySchema = createInsertSchema(categoriesTable);
+export const POSTCategorySchema = z.object({
+  name: varcharSchema("name"),
+  description: textSchema("Description").nullish(),
+  is_active: isActiveSchema,
+});
 export const PUTCategorySchema = createUpdateSchema(categoriesTable);
 
 export const GETCategoriesQuerySchema = z
